Make lints into plugins
