+++ setup.cfg
@@ -14,6 +14,7 @@ classifiers =
 packages =
     ethereum_spec_tools
     ethereum_spec_tools/lint
+    ethereum_spec_tools/lint/lints
     ethereum
     ethereum/frontier
     ethereum/frontier/utils
