+++ tox.ini
@@ -10,7 +10,7 @@ commands =
     isort src tests setup.py --check --diff --skip-glob "tests/fixtures/*"
     black src tests setup.py --check --diff --exclude "tests/fixtures/*"
     flake8 src tests setup.py
-    mypy src tests setup.py --exclude "tests/fixtures/*"
+    mypy src tests setup.py --exclude "tests/fixtures/*" --namespace-packages
     pytest --cov=ethereum --cov-report=term --cov-report "xml:{toxworkdir}/coverage.xml" --ignore-glob='tests/fixtures/*' -n 2
     ethereum-spec-lint
